Relocated `AutoscalingConfig` under ray.data.context
