+++ python/ray/data/context.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 import ray
-from ray._private.ray_constants import env_bool, env_integer
+from ray._private.ray_constants import env_bool, env_integer, env_float
 from ray._private.worker import WORKER_MODE
 from ray.data._internal.execution.autoscaler import AutoscalingConfig
 from ray.data._internal.logging import update_dataset_logger_for_worker
@@ -210,6 +210,47 @@ DEFAULT_MIN_HASH_SHUFFLE_AGGREGATOR_WAIT_TIME_IN_S = env_integer(
 DEFAULT_HASH_SHUFFLE_AGGREGATOR_HEALTH_WARNING_INTERVAL_S = env_integer(
     "RAY_DATA_HASH_SHUFFLE_AGGREGATOR_HEALTH_WARNING_INTERVAL_S", 30
 )
+
+
+DEFAULT_ACTOR_POOL_UTIL_UPSCALING_THRESHOLD: float = env_float(
+    "RAY_DATA_DEFAULT_ACTOR_POOL_UTIL_UPSCALING_THRESHOLD",
+    1.0,
+)
+
+DEFAULT_ACTOR_POOL_UTIL_DOWNSCALING_THRESHOLD: float = env_float(
+    "RAY_DATA_DEFAULT_ACTOR_POOL_UTIL_DOWNSCALING_THRESHOLD",
+    0.5,
+)
+
+
+@DeveloperAPI
+@dataclass
+class AutoscalingConfig:
+    # Actor Pool utilization threshold for upscaling. Once Actor Pool
+    # exceeds this utilization threshold it will start adding new actors.
+    #
+    # NOTE: Actor Pool utilization is defined as ratio of
+    #
+    #   - Number of submitted tasks to
+    #   - Max number of tasks the current set of Actors in the pool could run
+    #     (defined as Ray Actor's `max_concurrency` * `pool.num_running_actors`)
+    #
+    # This utilization value could exceed 100%, when the number of submitted tasks
+    # exceed available concurrency-slots to run them in the current set of actors.
+    #
+    # This is possible when `max_tasks_in_flight_per_actor` (defaults to 2 x
+    # of `max_concurrency`) > Actor's `max_concurrency` and allows to overlap
+    # task execution with the fetching of the blocks for the next task providing
+    # for ability to negotiate a trade-off between autoscaling speed and resource
+    # efficiency (ie making tasks wait instead of immediately triggering execution)
+    actor_pool_util_upscaling_threshold: float = (
+        DEFAULT_ACTOR_POOL_UTIL_UPSCALING_THRESHOLD,
+    )
+
+    # Actor Pool utilization threshold for downscaling
+    actor_pool_util_downscaling_threshold: float = (
+        DEFAULT_ACTOR_POOL_UTIL_DOWNSCALING_THRESHOLD,
+    )
 
 
 def _execution_options_factory() -> "ExecutionOptions":
